+++ src/main/scala/tscfg/gen4tests.scala
@@ -65,9 +65,6 @@ object gen4tests {
       .collect { case linePat(xs) => xs.trim }
       .flatMap(_.split("\\s+"))
       .map {
-        case opt @ "--scala:2.12" =>
-          OptFromFile(opt, Some("scala"))
-
         case opt @ "--scala:bt" =>
           OptFromFile(opt, Some("scala"))
 
@@ -107,8 +104,6 @@ object gen4tests {
     val baseGenOpts: GenOpts = {
       var genOpts = GenOpts("tscfg.example", "?")
       optsFromFile foreach {
-        case "--scala:2.12" => genOpts = genOpts.copy(s12 = true)
-
         case "--scala:bt" => genOpts = genOpts.copy(useBackticks = true)
 
         case "--java:getters" => genOpts = genOpts.copy(genGetters = true)