+++ src/main/scala/tscfg/generators/java/JavaGen.scala
@@ -632,7 +632,6 @@ object JavaGen {
   def generate(
       filename: String,
       showOut: Boolean = false,
-      j7: Boolean = false,
       assumeAllRequired: Boolean = false,
       genGetters: Boolean = false,
       genRecords: Boolean = false,
@@ -674,7 +673,6 @@ object JavaGen {
     val genOpts = GenOpts(
       "tscfg.example",
       className,
-      j7 = j7,
       genGetters = genGetters,
       genRecords = genRecords,
       useOptionals = useOptionals,